frontend/requirements.txt
FLASK==1.1.1
keras
tensorflow
efficientnet
Pillow
